+++ features/step_definitions/custom_web_steps.rb
@@ -190,13 +190,34 @@ end
 And /^I scroll down$/ do
   page.execute_script("window.scrollBy(0,3000000)")
 end
+And /^I scroll down on the notifications dropdown$/ do
+  page.execute_script("$('.notifications').scrollTop(350)")
+end
 
 Then /^I should have scrolled down$/ do
   page.evaluate_script("window.pageYOffset").should > 0
 end
 
+Then /^I should have scrolled down on the notification dropdown$/ do
+  page.evaluate_script("$('.notifications').scrollTop()").should > 0
+end
+
+
 Then /^the notification dropdown should be visible$/ do
   find(:css, "#notification_dropdown").should be_visible
+end
+
+Then /^the notification dropdown scrollbar should be visible$/ do
+  find(:css, ".ps-active-y").should be_visible
+end
+
+Then /^there should be (\d+) notifications loaded$/ do |n|
+  result = page.evaluate_script("$('.notification_element').length")
+  result.should == n.to_i
+end
+
+And "I wait for notifications to load" do
+  page.should_not have_selector(".loading")
 end
 
 When /^I resize my window to 800x600$/ do